Add config-option to receive even self-sent messages.
The config option is at { "settings" : { "receive-while-sending": 0 or 1 } }
and the pilight-daemon is using it.  The purpose of this option is to test
protocol handler implementations like this:
Run pilight-receive at term one to see daemon logs. Then running a

  pilight-send -p proto --id xx --unit yy etc...

at term two should produce a sent and a receive message stating the same
id, unit and command codes. If they differ constantly the implementation
should be reviewed.

Note a strange thing: due to thread priorities, the receiving-log-message
may sometime appear before the send-log-message.